requirements.txt
Django==3.2.7
loguru==0.5.3
lxml==4.6.3
pandas==1.3.3
django-bootstrap4==3.0.1
django-bootstrap-datepicker-plus==3.0.5
gunicorn==20.0.4
